+++ Release Notes.md
@@ -3,6 +3,14 @@
 I will bump revision by revision, until SwiftKit has all functionality that it should have from iExtra. I will then bump it to `1.0.0`.
 
 
+## 0.5.1
+
+### ✨ New features
+
+* `String+Slugify` is a new extension that can convert a string to a slugified version.
+
+
+
 ## 0.5.0
 
 This version adjusts code that made the demo not being able to use the SPM package instead of the source files.